requirements.txt
# 接口监控脚本依赖库
# Python版本要求: >=3.7

# HTTP请求库
requests>=2.31.0

# 定时任务库
schedule>=1.2.0

# YAML解析库
pyyaml>=6.0

# 文件监控（配置热更新）
watchdog>=3.0.0

# 日志轮转
loguru>=0.7.0

# 单元测试
pytest>=7.4.0
pytest-cov>=4.1.0

# 代码格式化
black>=23.0.0

# 代码检查
pylint>=2.17.0
